rename results page
