+++ src/routers/forumQuestions.ts
@@ -2,7 +2,7 @@ import type { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
 import { default as express } from 'express';
 import type * as db from '../db';
 import { getQuestionHearts } from '../services/forumQuestions';
-import { getResultStatistics } from '../services/resultsPage';
+import { getResultStatistics } from '../services/statistics';
 import { isLoggedIn } from '../middleware/isLoggedIn';
 const router = express.Router();
