+++ src/services/statistics.ts
